calm-simple/requirements.txt
streamlit>=1.31.0
websocket-client>=1.6.0
